refactor token members into the header
